+++ PeacefulPaths/src/main/java/com/brogramer/peacefulPaths/entity/TherapistNotesHistory.java
@@ -11,7 +11,7 @@ public class TherapistNotesHistory {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
 
     @ManyToOne
     @JoinColumn(name = "notes_id")
@@ -29,11 +29,11 @@ public class TherapistNotesHistory {
 
     public TherapistNotesHistory() {}
 
-    public TherapistNotesHistory(Long id) {
+    public TherapistNotesHistory(int id) {
         this.id = id;
     }
 
-    public TherapistNotesHistory(Long id, Notes notes, int clientId, int therapistId, LocalDateTime dateAdded) {
+    public TherapistNotesHistory(int id, Notes notes, int clientId, int therapistId, LocalDateTime dateAdded) {
         this.id = id;
         this.notes = notes;
         this.clientId = clientId;
@@ -41,11 +41,11 @@ public class TherapistNotesHistory {
         this.dateAdded = dateAdded;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
